Final commit (11/11/2018):
Most of the work on the final day was tweaking presentation of the assignment, and ensuring that all texts were referenced. It is made clear that the descriptions of the six probability distribution functions come from wikipedia.

An attempt to overlay each histogram with a smooth continuous distribution created with scipy was abandoned despite earlier success plotting a normal distribution this way.